refactor: `Detail 페이지` 로직 커스텀훅으로 분리
